requirements.txt
fastapi>=0.124.4
numpy>=2.0.2
pillow>=12.0.0
python-multipart>=0.0.20
tensorflow-cpu<2.20
uvicorn>=0.38.0